requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
PyGithub==1.35
PyJWT==1.5.2
python-dotenv==0.6.5
python-frontmatter==0.4.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
tornado==4.5.2
urllib3==1.22
